ai-nutrition-analyzer/ai-services/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
requests==2.31.0
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.4.0
pandas==2.1.4
scipy==1.11.4
httpx==0.26.0
pydub==0.25.1
speechrecognition==3.10.1
openai-whisper==20231117
